Break up bootstrap script
It was always a pain to comment out certain sections of the bootstrap script when tweaking/modifying part of it.

Support running a specific step in the bootstrap script for re-run after a tweak or a quick test.
Passing no arguments will run all steps

The main script makes use of bash default variable assignment
https://bash.cyberciti.biz/guide/Default_shell_variables_value
and array loop syntax:
https://www.cyberciti.biz/faq/bash-for-loop-array/